+++ test/integration/save.js
@@ -66,4 +66,56 @@ describe("INTEGRATION - SAVE", function () {
 			done();
 		});
 	});
+	
+	it("model.save() with passed in object should save and add _id to the instance", function (done) {
+		var Model = Platos.create("Model");
+		
+		var instance = new Model({ test: "property" });
+
+		instance.save({ test2: "property" }, function (err, document) {
+			instance.should.have.property("_id");
+			instance.should.have.property("test");
+			instance.should.not.have.property("test2");
+			document.should.not.have.property("test");
+			document.should.have.property("test2")
+			
+			done();
+		});
+	});
+	
+	it("model.save() with pre hook that mutates the object should save and add _id to the Model instance", function (done) {
+		var Model = Platos.create("Model");
+		
+		Model.pre("save", function (next, object, callback) {
+			next({ test: "replace" }, callback);
+		});
+		
+		var instance = new Model();
+		
+		instance.save(function (err, document) {
+			instance.should.have.property("_id");
+			instance.should.not.have.property("test");
+			document.should.have.property("test");
+			
+			done();
+		});
+	});
+	
+	it("model.save() with empty pre hook should save and add _id to the instance", function (done) {
+		var Model = Platos.create("Model");
+		
+		Model.pre("save", function (next, object, callback) {
+			next(object, callback);
+		});
+		
+		var instance = new Model({ test: "property" });
+		
+		instance.save(function (err, document) {
+			instance.should.have.property("_id");
+			instance.should.have.property("test");
+			document.should.have.property("test");
+			
+			done();
+		});
+	});
 });